Added authorization using gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
@@ -58,7 +59,14 @@ class AttendeeController extends Controller
      */
     public function destroy(Event $event, Attendee $attendee)
     {
-        $attendee->delete();
-        return response(status: 204);
+        try {
+            $this->authorize("attendee-authorization", [$event, $attendee]);
+            $attendee->delete();
+            return response(status: 204);
+        }catch (AuthorizationException $exception) {
+            return response()->json([
+                "message" => $exception->getMessage()
+            ], 403);
+        }
     }
 }
